fix : category id 타입 변환
Long -> Byte

+++ backend/clothes/src/main/java/com/ssafy/ozz/clothes/clothes/dto/request/SearchCondition.java
@@ -1,6 +1,6 @@
 package com.ssafy.ozz.clothes.clothes.dto.request;
 
 public record SearchCondition (
-    Long categoryHighId,
-    Long categoryLowId
+    Byte categoryHighId,
+    Byte categoryLowId
 ){ }
